fix: rename to transformDots
