+++ dotcom-rendering/src/model/enhanceStandfirst.ts
@@ -1,3 +1,3 @@
-import { transformDots } from './transforDots';
+import { transformDots } from './transformDots';
 
 export const enhanceStandfirst = (html: string): string => transformDots(html);
