+++ dotcom-rendering/src/model/transformDots.ts
